05-11-20 conclusão de testes unitários de cenários ideais. inserção de exceções em aluno, mentor e mentoria.

+++ demo/src/main/java/com/example/demo/service/MentorService.java
@@ -2,6 +2,7 @@ package com.example.demo.service;
 
 import com.example.demo.dto.MentorDTO;
 import com.example.demo.dto.mapper.MentorMapper;
+import com.example.demo.exceptions.BadRequestException;
 import com.example.demo.exceptions.NotFoundException;
 import com.example.demo.model.Mentor;
 import com.example.demo.repository.MentorRepository;
@@ -46,46 +47,52 @@ public class MentorService {
     }
 
     public Optional<MentorDTO> getMentorByIndex(Long id)/*throws Exception*/{
-
 //        Optional<MentorDTO> mentorDTO = mentorRepository.findById(id).map(MentorMapper::toMentorDTO);
 //        if (!mentorDTO.isPresent()){
 //            throw new Exception ("id "+id+" não encontrado.");
 //        }
 //        return mentorDTO.get();
-        if (mentorRepository.findById(id).get().getActive()){
+        if (mentorRepository.findById(id).isPresent() && mentorRepository.findById(id).get().getActive()){
             return Optional.of(mentorRepository.findById(id)
-                    .map(mentorMapper::toMentorDTO))
-                    .orElse(Optional.empty());
+                    .map(mentorMapper::toMentorDTO)).get();
         } else{
-            return Optional.empty();
+            throw new NotFoundException("Não há mentor ativo com o ID informado.");
         }
     }
 
-    public Optional<MentorDTO> criaMentor(MentorDTO mentorDTO)/* throws Exception*/{
+    public Optional<MentorDTO> criaMentor(MentorDTO mentorDTO){
 
-        mentorDTO.setActive(true);
-        return Optional.of(mentorMapper
-                .toMentorDTO(mentorRepository.save(mentorMapper.toMentor(mentorDTO))));
+        if((mentorDTO.getName().length()>=3 && mentorDTO.getName().length()<=50) && mentorDTO.getPais().length()<=50 && mentorDTO.getEscola().length()<=50){
+            mentorDTO.setActive(true);
+            return Optional.of(mentorMapper
+                    .toMentorDTO(mentorRepository.save(mentorMapper.toMentor(mentorDTO))));
+        }else{
+            throw new BadRequestException("Valores inválidos informados.");
+        }
+
     }
 
-    public Optional<MentorDTO> removeMentorByIndex(Long id){
+    public Optional<MentorDTO> removeMentorByIndex(Long id) {
         Optional<Mentor> mentor = mentorRepository.findById(id);
 
-        if(mentor.isPresent()) {
+        if((mentor.isPresent()) && (mentor.get().getActive())) {
             mentor.get().setActive(false);
             mentoriaService.setActiveByMentor(false, id);
             return Optional.of(mentorMapper.toMentorDTO(mentorRepository.save(mentor.get())));// mentorRepository.deleteById(id);
         }else{
-            throw new NotFoundException();
+            throw new NotFoundException("Não há mentor ativo com o ID informado.");
         }
 
     }
 
     public Optional<MentorDTO> alteraMentor(Long id, MentorDTO mentorDTO){
         Optional<Mentor> mentor = mentorRepository.findById(id);
-        if (!mentor.isPresent()){
-            return Optional.empty();
+        if ((!mentor.isPresent()) || (!mentor.get().getActive())){
+            throw new NotFoundException("Não há mentor ativo com o ID informado.");
         }else{
+            if(mentorDTO.getName().length()<3 || mentorDTO.getName().length()>50 || mentorDTO.getPais().length()>50 || mentorDTO.getEscola().length()>50){
+                throw new BadRequestException("Valores inválidos informados.");
+            }
             mentorDTO.setId(id);
             mentorDTO.setActive(true);
             return Optional.of(mentorMapper
@@ -94,10 +101,12 @@ public class MentorService {
     }
     public Optional<MentorDTO> reativaMentor(Long id){
         Optional<Mentor> mentor = mentorRepository.findById(id);
-        if(mentor.isPresent()) {
+        if((mentor.isPresent()) && (!mentor.get().getActive())) {
             mentor.get().setActive(true);
+            return Optional.of(mentorMapper.toMentorDTO(mentorRepository.save(mentor.get())));
+        }else {
+            throw new NotFoundException("Não há mentor inativo com o ID informado.");
         }
-        return Optional.of(mentorMapper.toMentorDTO(mentorRepository.save(mentor.get())));
     }
     }
